# backend/src/main/java/com/company/cms/shared/api/ErrorResponse.java
package com.company.cms.shared.api;

import java.time.Instant;
import java.util.List;

public record ErrorResponse(
        String code,
        String message,
        List<String> details,
        Instant timestamp
) {
    public static ErrorResponse of(String code, String message) {
        return new ErrorResponse(code, message, List.of(), Instant.now());
    }

    public static ErrorResponse of(String code, String message, List<String> details) {
        return new ErrorResponse(code, message, details == null ? List.of() : details, Instant.now());
    }
}