# backend/src/main/resources/openapi/cms-api.yaml
openapi: 3.0.3
info:
  title: CMS Internal Knowledge Portal API
  version: 0.1.0
  description: 사내 지식·문서·공지 CMS와 사용자 포털을 위한 REST API 계약
servers:
  - url: https://cms.example.com
    description: production
  - url: http://localhost:8080
    description: local
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/content:
    get:
      tags: [Content]
      operationId: listContent
      summary: CMS 콘텐츠 목록 조회
      parameters:
        - name: contentType
          in: query
          schema:
            $ref: "#/components/schemas/ContentType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ContentStatus"
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          schema:
            type: string
            maxLength: 200
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 콘텐츠 목록
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContentSummary"
                  page:
                    $ref: "#/components/schemas/PageInfo"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Content]
      operationId: createContent
      summary: 콘텐츠 초안 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentCreateRequest"
      responses:
        "201":
          description: 생성된 콘텐츠
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/content/{contentId}:
    parameters:
      - $ref: "#/components/parameters/ContentId"
    get:
      tags: [Content]
      operationId: getContentForAdmin
      summary: CMS 콘텐츠 상세 조회
      responses:
        "200":
          description: 콘텐츠 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Content]
      operationId: updateContent
      summary: 콘텐츠 초안 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdateRequest"
      responses:
        "200":
          description: 수정된 콘텐츠
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: 최신 버전 충돌
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Content]
      operationId: deleteDraftContent
      summary: 발행 전 초안 삭제
      responses:
        "204":
          description: 삭제됨
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/content/{contentId}/versions:
    parameters:
      - $ref: "#/components/parameters/ContentId"
    get:
      tags: [Content]
      operationId: listContentVersions
      summary: 콘텐츠 버전 이력 조회
      responses:
        "200":
          description: 버전 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentVersion"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/content/{contentId}/submit:
    parameters:
      - $ref: "#/components/parameters/ContentId"
    post:
      tags: [Content]
      operationId: submitContentForReview
      summary: 콘텐츠 검토 요청
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReviewRequest"
      responses:
        "202":
          description: 검토 요청됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalTask"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/content/{contentId}/approval-decisions:
    parameters:
      - $ref: "#/components/parameters/ContentId"
    post:
      tags: [Content]
      operationId: decideContentApproval
      summary: 콘텐츠 승인 또는 반려
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: 승인 작업 처리 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalTask"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/content/{contentId}/publish:
    parameters:
      - $ref: "#/components/parameters/ContentId"
    post:
      tags: [Content]
      operationId: publishContent
      summary: 콘텐츠 즉시 또는 예약 발행
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
      responses:
        "200":
          description: 발행 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/content/{contentId}/archive:
    parameters:
      - $ref: "#/components/parameters/ContentId"
    post:
      tags: [Content]
      operationId: archiveContent
      summary: 콘텐츠 보관 처리
      responses:
        "200":
          description: 보관된 콘텐츠
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/portal/content/{contentId}:
    parameters:
      - $ref: "#/components/parameters/ContentId"
    get:
      tags: [Portal]
      operationId: getPortalContent
      summary: 포털 콘텐츠 상세 조회
      responses:
        "200":
          description: 공개 대상 검증 후 콘텐츠 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/admin/categories:
    get:
      tags: [Admin]
      operationId: listCategories
      summary: 카테고리 목록 조회
      responses:
        "200":
          description: 카테고리 계층 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      tags: [Admin]
      operationId: createCategory
      summary: 카테고리 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
      responses:
        "201":
          description: 생성된 카테고리
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/tags:
    get:
      tags: [Admin]
      operationId: listTags
      summary: 태그 목록 조회
      responses:
        "200":
          description: 태그 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      tags: [Admin]
      operationId: createTag
      summary: 태그 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagRequest"
      responses:
        "201":
          description: 생성된 태그
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
...
  /api/v1/admin/analytics/contents:
    get:
      tags: [Admin]
      operationId: getContentAnalytics
      summary: 콘텐츠 운영 통계 조회
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: contentType
          in: query
          schema:
            $ref: "#/components/schemas/ContentType"
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 콘텐츠 운영 지표
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ContentId:
      name: contentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: 요청 값 검증 실패
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: 권한 부족
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ContentType:
      type: string
      enum: [ARTICLE, DOCUMENT, NOTICE]
    ContentStatus:
      type: string
      enum: [DRAFT, IN_REVIEW, REJECTED, APPROVED, SCHEDULED, PUBLISHED, ARCHIVED, EXPIRED]
    RoleCode:
      type: string
      enum: [ADMIN, EDITOR, REVIEWER, EMPLOYEE, VIEWER]
    UserProfile:
      type: object
      required: [id, email, displayName, roles, permissions]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        department:
          $ref: "#/components/schemas/Department"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleCode"
        permissions:
          type: array
          items:
            type: string
    Department:
      type: object
      required: [id, name, isActive]
      properties:
        id:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        code:
          type: string
        name:
          type: string
        isActive:
          type: boolean
    Category:
      type: object
      required: [id, name, slug, isActive]
      properties:
        id:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        sortOrder:
          type: integer
        isActive:
          type: boolean
    CategoryRequest:
      type: object
      required: [name, slug]
      properties:
        parentId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          maxLength: 80
        slug:
          type: string
          maxLength: 120
        description:
          type: string
        sortOrder:
          type: integer
          minimum: 0
    Tag:
      type: object
      required: [id, name, isActive]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        isActive:
          type: boolean
    TagRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 40
    AudienceRule:
      type: object
      required: [audienceType]
      properties:
        audienceType:
          type: string
          enum: [ALL_EMPLOYEES, DEPARTMENT, ROLE, USER]
        audienceRefId:
          type: string
          nullable: true
          description: "`ALL_EMPLOYEES` 외 대상 참조 식별자"
    Attachment:
      type: object
      required: [id, fileName, mimeType, sizeBytes, scanStatus]
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        scanStatus:
          type: string
          enum: [PENDING, CLEAN, INFECTED, FAILED]
        downloadUrl:
          type: string
          format: uri
          nullable: true
    ContentSummary:
      type: object
      required: [id, contentType, status, title, summary, category, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        contentType:
          $ref: "#/components/schemas/ContentType"
        status:
          $ref: "#/components/schemas/ContentStatus"
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        isImportant:
          type: boolean
        requiresAcknowledgement:
          type: boolean
        publishedAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
    ContentDetail:
      allOf:
        - $ref: "#/components/schemas/ContentSummary"
        - type: object
          required: [body, author, audiences, versionNumber]
          properties:
            body:
              type: string
              description: 리치 텍스트 본문 payload
            author:
              $ref: "#/components/schemas/UserProfile"
            audiences:
              type: array
              items:
                $ref: "#/components/schemas/AudienceRule"
            attachments:
              type: array
              items:
                $ref: "#/components/schemas/Attachment"
            versionNumber:
              type: integer
            scheduledAt:
              type: string
              format: date-time
              nullable: true
            expiresAt:
              type: string
              format: date-time
              nullable: true
    ContentCreateRequest:
      type: object
      required: [contentType, title, body, categoryId, audiences]
      properties:
        contentType:
          $ref: "#/components/schemas/ContentType"
        title:
          type: string
          maxLength: 150
        summary:
          type: string
          maxLength: 300
        body:
          type: string
        categoryId:
          type: string
          format: uuid
        tagIds:
          type: array
          items:
            type: string
            format: uuid
        audiences:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AudienceRule"
        isImportant:
          type: boolean
          default: false
        requiresAcknowledgement:
          type: boolean
          default: false
        expiresAt:
          type: string
          format: date-time
          nullable: true
    ContentUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/ContentCreateRequest"
        - type: object
          required: [baseVersionNumber, changeSummary]
          properties:
            baseVersionNumber:
              type: integer
            changeSummary:
              type: string
              maxLength: 500
    ContentVersion:
      type: object
      required: [id, versionNumber, titleSnapshot, changeSummary, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        versionNumber:
          type: integer
        titleSnapshot:
          type: string
        summarySnapshot:
          type: string
        changeSummary:
          type: string
        createdBy:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string
          format: date-time
    SubmitReviewRequest:
      type: object
      required: [reviewerId, changeSummary]
      properties:
        reviewerId:
          type: string
          format: uuid
        changeSummary:
          type: string
          maxLength: 500
    ApprovalDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [APPROVE, REJECT]
        decisionComment:
          type: string
          maxLength: 1000
    ApprovalTask:
      type: object
      required: [id, contentId, reviewerId, status, requestedAt]
      properties:
        id:
          type: string
          format: uuid
        contentId:
          type: string
          format: uuid
        reviewerId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, CANCELLED]
        decisionComment:
          type: string
          nullable: true
        requestedAt:
          type: string
          format: date-time
        decidedAt:
          type: string
          format: date-time
          nullable: true
    PublishRequest:
      type: object
      properties:
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
    Publication:
      type: object
      required: [id, contentId, publicationStatus]
      properties:
        id:
          type: string
          format: uuid
        contentId:
          type: string
          format: uuid
        publicationStatus:
          type: string
          enum: [SCHEDULED, PUBLISHED, ARCHIVED, EXPIRED]
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
...
    AnalyticsResponse:
      type: object
      required: [period, metrics, popularContent, topSearchQueries, noResultQueries, staleContent]
      properties:
        period:
          type: object
          required: [from, to]
          properties:
            from:
              type: string
              format: date
            to:
              type: string
              format: date
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/MetricValue"
        popularContent:
          type: array
          items:
            $ref: "#/components/schemas/ContentSummary"
        topSearchQueries:
          type: array
          items:
            $ref: "#/components/schemas/SearchQueryMetric"
        noResultQueries:
          type: array
          items:
            $ref: "#/components/schemas/SearchQueryMetric"
        staleContent:
          type: array
          items:
            $ref: "#/components/schemas/ContentSummary"
    MetricValue:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: number
          format: double
        unit:
          type: string
    SearchQueryMetric:
      type: object
      required: [queryText, count, resultCount]
      properties:
        queryText:
          type: string
        count:
          type: integer
        resultCount:
          type: integer
    PageInfo:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string